+++ app/src/main/java/com/essensol/techmeq/OnSelectedListener.java
@@ -2,6 +2,7 @@ package com.essensol.techmeq;
 
 import com.essensol.techmeq.Model.mProductModel;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface OnSelectedListener {
@@ -10,5 +11,5 @@ public interface OnSelectedListener {
 //    void getProductDetails(List<mProductModel> items);
 
 
-    void getProductDetails(int Product_Id,int ProductCatId,double TaxPercent,String ProductName,double Sales_Price,boolean Status);
+    void getProductDetails(int Product_Id, int ProductCatId, BigDecimal TaxPercent, String ProductName, BigDecimal Sales_Price, boolean Status);
 }
